move sim_total_artificial_heart class to seperate file
